requirements.txt
streamlit==1.32.0
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1
matplotlib==3.8.3
plotly==5.20.0
flask==3.0.2
paho-mqtt==1.6.1
python-pptx==0.6.23
joblib==1.3.2
tensorflow==2.15.0
sqlalchemy==2.0.28
